+++ src/Schema/Directives/ComplexityDirective.php
@@ -59,9 +59,8 @@ GRAPHQL;
 
         return
             // Default complexity for this field itself
-            // TODO add this in v6
-            // 1 +
+            1
             // Scale children complexity by the expected number of results
-            $childrenComplexity * $expectedNumberOfChildren;
+            + $childrenComplexity * $expectedNumberOfChildren;
     }
 }